requirements.txt
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
certifi==2025.4.26
chardet==5.2.0
click==8.1.8
crispy-bootstrap5==2025.4
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
future==1.0.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jedi==0.19.2
packaging==25.0
parso==0.8.4
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
ptpython==3.0.30
Pygments==2.19.1
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
textfsm==2.1.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.9.0
